add device search to device.go
